# server/openapi.yaml
openapi: 3.0.3
info:
  title: AgentPay API
  description: |
    AgentPay is a decentralized compute marketplace where AI agents pay for
    GPU/inference/data services via x402 micropayments on Movement.

    ## Authentication

    All paid endpoints require x402 payment. When you make a request without payment,
    you'll receive a `402 Payment Required` response with pricing information.

    To pay, include the `X-PAYMENT` header with a signed payment payload.

    ## Pricing

    Prices are returned in octas (1 MOVE = 10^8 octas). The pricing formula is:
    `Customer pays = Provider cost + 15% margin`
  version: 0.1.0
  contact:
    name: AgentPay Support
    url: https://agentpay.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4402
    description: Local development server
  - url: https://api.agentpay.io
    description: Production server
...
paths:
...
  /v1/usage:
    get:
      tags: [Usage]
      summary: Get usage statistics
      description: Get usage statistics and spending history
      parameters:
        - name: payer
          in: query
          schema:
            type: string
          description: Filter by payer wallet address
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Number of days to include
      responses:
        '200':
          description: Usage statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageStats'
  /v1/usage/history:
    get:
      tags: [Usage]
      summary: Get payment history
      description: Get payment history for a specific payer
      parameters:
        - name: payer
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  total:
                    type: integer
...
components:
...
  schemas:
...
    UsageStats:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalRequests:
              type: integer
            totalSpent:
              type: string
            averageCost:
              type: string
            currency:
              type: string
        daily:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              requests:
                type: integer
              amount:
                type: string
    Payment:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: string
        serviceType:
          type: string
        model:
          type: string
        provider:
          type: string
        txHash:
          type: string
        timestamp:
          type: string
          format: date-time
...